Centered on current location

+++ app/src/main/java/com/whiterabbit/pisabike/Constants.java
@@ -2,6 +2,6 @@ package com.whiterabbit.pisabike;
 
 
 public class Constants {
-    public static final float MY_LATITUDE = 43.7228f;
-    public static final float MY_LONGITUDE = 10.4017f;
+    public static final double MY_LATITUDE = 43.7228;
+    public static final double MY_LONGITUDE = 10.4017;
 }
